requirements.txt
streamlit
tensorflow
numpy
pillow
pandas
matplotlib
seaborn
